LRQA-19505 move
