+++ modules/test/jenkins-results-parser/src/test/resources/com/liferay/results/parser/load/balancer/dependencies/LoadBalanceUtilTest/test-1/test-1-19/computer/api/json?pretty&tree=computer[displayName,idle,offline]
