+++ modules/test/jenkins-results-parser/src/test/resources/com/liferay/results/parser/load/balancer/dependencies/LoadBalanceUtilTest/test-2/test-2-1/computer/api/json?pretty&tree=computer[displayName,idle,offline]
